removed API types from various handlers

+++ src/SFA.DAS.Commitments.Application/Commands/UpdateApprenticeshipStatus/UpdateApprenticeshipStatusValidator.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using FluentValidation;
-using SFA.DAS.Commitments.Api.Types.Apprenticeship.Types;
+
+using SFA.DAS.Commitments.Domain.Entities;
 
 namespace SFA.DAS.Commitments.Application.Commands.UpdateApprenticeshipStatus
 {
@@ -11,7 +12,7 @@ namespace SFA.DAS.Commitments.Application.Commands.UpdateApprenticeshipStatus
             RuleFor(x => x.AccountId).GreaterThan(0);
             RuleFor(x => x.ApprenticeshipId).GreaterThan(0);
             RuleFor(x => x.PaymentStatus).NotNull()
-                .DependentRules(y => y.RuleFor(z => z.PaymentStatus).Must(a => Enum.IsDefined(typeof(PaymentStatus), (short)a)));
+                .DependentRules(y => y.RuleFor(z => z.PaymentStatus).Must(a => Enum.IsDefined(typeof(PaymentStatus), (int)a)));
         }
     }
 }
